Move stats/git.py -> git.py
